Move boto3 tests to separate file
